Moved File.getFolderIcon() to Extensions.kt
