+++ app/src/main/java/com/infomaniak/drive/data/models/File.kt
@@ -18,11 +18,9 @@
 package com.infomaniak.drive.data.models
 
 import android.content.Context
-import android.graphics.drawable.Drawable
 import android.net.Uri
 import android.os.Parcelable
 import android.webkit.MimeTypeMap
-import androidx.core.content.ContextCompat
 import androidx.core.content.FileProvider
 import androidx.work.WorkInfo
 import androidx.work.WorkManager
@@ -38,7 +36,6 @@ import com.infomaniak.drive.utils.RealmListParceler.FileRealmListParceler
 import com.infomaniak.drive.utils.RealmListParceler.IntRealmListParceler
 import com.infomaniak.drive.utils.Utils.INDETERMINATE_PROGRESS
 import com.infomaniak.drive.utils.Utils.ROOT_ID
-import com.infomaniak.drive.utils.getTintedDrawable
 import com.infomaniak.lib.core.BuildConfig
 import io.realm.RealmList
 import io.realm.RealmObject
@@ -427,6 +424,13 @@ open class File(
     }
 
     companion object {
+
+        /**
+         * This method is here, and not directly a class method in the File class, because of a supposed Realm bug.
+         * When we try to put it in the File class, the app doesn't build anymore, because of a "broken method".
+         * This is not the only method in this case, search this comment in the project, and you'll see.
+         * Realm's Github issue: https://github.com/realm/realm-java/issues/7637
+         */
         fun File.getCloudAndFileUris(context: Context, userDrive: UserDrive = UserDrive()): Pair<Uri, Uri> {
             val cloudUri = CloudStorageProvider.createShareFileUri(context, this, userDrive)!!
             val offlineFile = getOfflineFile(context, userDrive.userId)
@@ -436,29 +440,15 @@ open class File(
             } else cloudUri
         }
 
+        /**
+         * This method is here, and not directly a class method in the File class, because of a supposed Realm bug.
+         * When we try to put it in the File class, the app doesn't build anymore, because of a "broken method".
+         * This is not the only method in this case, search this comment in the project, and you'll see.
+         * Realm's Github issue: https://github.com/realm/realm-java/issues/7637
+         */
         fun getOfflineFolder(context: Context): java.io.File {
             val mediaFolder = context.externalMediaDirs?.firstOrNull() ?: context.filesDir
             return java.io.File(mediaFolder, context.getString(R.string.EXPOSED_OFFLINE_DIR))
-        }
-
-        fun File.getFolderIcon(context: Context): Drawable? {
-            return if (isFolder()) {
-                val icon: Any? = when (getVisibilityType()) {
-                    VisibilityType.IS_TEAM_SPACE -> R.drawable.ic_folder_common_documents
-                    VisibilityType.IS_SHARED_SPACE -> R.drawable.ic_folder_shared
-                    VisibilityType.IS_COLLABORATIVE_FOLDER -> context.getTintedDrawable(R.drawable.ic_folder_dropbox, color)
-                    else -> {
-                        if (isDisabled()) {
-                            R.drawable.ic_folder_disable
-                        } else {
-                            context.getTintedDrawable(R.drawable.ic_folder_filled, color)
-                        }
-                    }
-                }
-                if (icon is Int) ContextCompat.getDrawable(context, icon) else (icon as Drawable)
-            } else {
-                null
-            }
         }
     }
 }
